Add Naive and KDTree tracers support.
A tracer is a class that is able to trace a ray and find where it hits an object of the scene.
Naive tracer is the naive way of tracing a ray: check for each object whether it's hit by the ray.
A KDTree tracer implements a k-d tree data structure that allows for sub-linear tracing; this is done by sub-dividing the scene box using multiple axis-aligned planes.